Fixed Transfer when Adjusted Cost is set
Added configuration to allow a dev to set a fake user instead of using the authorization token from the auth service

Related work items: #29693

+++ IRMA/Projects/IrmaMobile/IrmaMobile/Domain/Models/TransferOrderItemModel.cs
@@ -10,7 +10,7 @@ namespace IrmaMobile.Domain.Models
         public int QuantityOrdered { get; set; }
         public int ItemKey { get; set; }
         public int QuantityUnit { get; set; }
-        public int AdjustedCost { get; set; }
+        public decimal AdjustedCost { get; set; }
         public int ReasonCodeDetailId { get; set; }
     }
 }
